Removed old bandersnatch_vrfs
