+++ src/ring_vrf_impl.rs
@@ -1,106 +1,86 @@
+#![allow(unused)]
+
 use alloc::vec;
 use core::ops::Range;
 
-use ark_scale::ArkScale;
+use ark_ec_vrfs::{ring::Verifier, suites::bandersnatch};
+use ark_scale::{impl_scale_via_ark, ArkScale, ArkScaleMaxEncodedLen};
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use bandersnatch_vrfs::ring::{KzgVk, RingCommitment, VerifierKey};
-use bandersnatch_vrfs::{bls12_381, deserialize_publickey, PUBLIC_KEY_LENGTH};
-use bandersnatch_vrfs::{
-	ring::ProverKey, zcash_consts, IntoVrfInput, Message, PublicKey as PublicKeyInternal,
-	RingVerifier, SecretKey, Transcript, VrfInput,
-};
-#[cfg(feature = "std")]
-use bandersnatch_vrfs::{ring::StaticProverKey, ring::KZG, RingProver};
+use scale_info::TypeInfo;
 
 use super::*;
 
-pub use bandersnatch_vrfs;
-
-type ThinVrfSignature = bandersnatch_vrfs::ThinVrfSignature<0>;
-type RingVrfSignature = bandersnatch_vrfs::RingVrfSignature<1>;
+#[cfg(feature = "std")]
+pub(crate) const PCS_PARAMS_ZCASH: &[u8] =
+	include_bytes!("ring-data/zcash-srs-2-16-uncompressed.bin");
 
 #[cfg(feature = "small-ring")]
-mod domain_params {
-	use super::*;
-	pub const DOMAIN_SIZE: usize = 1 << 9;
-	pub(crate) const EMPTY_RING: RingCommitment = zcash_consts::EMPTY_RING_ZCASH_9;
-	#[cfg(feature = "std")]
-	pub(crate) const OFFCHAIN_PK: &[u8] = include_bytes!("ring-data/zcash-9.pk");
-}
+pub const RING_SIZE: usize = 255; // TODO
 
 #[cfg(not(feature = "small-ring"))]
-mod domain_params {
-	use super::*;
-	pub const DOMAIN_SIZE: usize = 1 << 16;
-	pub(crate) const EMPTY_RING: RingCommitment = zcash_consts::EMPTY_RING_ZCASH_16;
-	#[cfg(feature = "std")]
-	pub(crate) const OFFCHAIN_PK: &[u8] = include_bytes!("ring-data/zcash-16.pk");
-}
-
-pub use domain_params::*;
-
-const THIN_SIGNATURE_CONTEXT: &[u8] = b"VerifiableBandersnatchThinSignature";
+pub const RING_SIZE: usize = 255; // TODO
 
 const VRF_INPUT_DOMAIN: &[u8] = b"VerifiableBandersnatchInput";
-const VRF_OUTPUT_DOMAIN: &[u8] = b"VerifiableBandersnatchInput";
+const THIN_SIGNATURE_CONTEXT: &[u8] = b"VerifiableBandersnatchThinSignature";
 
-const THIN_SIGNATURE_SIZE: usize = 65;
-const RING_SIGNATURE_SIZE: usize = 788;
+macro_rules! impl_scale {
+	($type_name:ident, $encoded_size:expr) => {
+		ark_scale::impl_scale_via_ark!($type_name);
+
+		impl scale::MaxEncodedLen for $type_name {
+			fn max_encoded_len() -> usize {
+				$encoded_size
+			}
+		}
+
+		impl scale_info::TypeInfo for $type_name {
+			type Identity = [u8; $encoded_size];
+			fn type_info() -> scale_info::Type {
+				Self::Identity::type_info()
+			}
+		}
+	};
+}
 
 #[cfg(feature = "std")]
-fn kzg() -> &'static KZG {
+fn ring_context() -> &'static bandersnatch::RingContext {
+	use ark_ec_vrfs::ring::PcsParams;
 	use std::sync::OnceLock;
-	static CELL: OnceLock<KZG> = OnceLock::new();
+	static CELL: OnceLock<bandersnatch::RingContext> = OnceLock::new();
 	CELL.get_or_init(|| {
-		let pk = StaticProverKey::deserialize_uncompressed_unchecked(OFFCHAIN_PK).unwrap();
-		KZG::kzg_setup(DOMAIN_SIZE, pk)
+		let pcs_params =
+			bandersnatch::PcsParams::deserialize_uncompressed_unchecked(PCS_PARAMS_ZCASH).unwrap();
+		bandersnatch::RingContext::from_srs(RING_SIZE, pcs_params).unwrap()
 	})
 }
 
-#[derive(Debug, Clone, Eq, PartialEq, CanonicalDeserialize, CanonicalSerialize)]
-pub struct MembersSet {
-	pub ring: RingCommitment,
-	pub kzg_raw_vk: KzgVk,
-}
+#[derive(Clone, CanonicalDeserialize, CanonicalSerialize)]
+pub struct MembersSet(bandersnatch::RingVerifierKeyBuilder);
 
-ark_scale::impl_scale_via_ark!(MembersSet);
-
+// TODO: check
 const MEMBERS_SET_SIZE: usize = 4 * 48 + 2 * 96 + 32 + 2 * 4; // 4 bls G1 + 2 bls G2 + jubjub + 2 usize
+impl_scale!(MembersSet, MEMBERS_SET_SIZE);
 
-impl scale_info::TypeInfo for MembersSet {
-	type Identity = [u8; MEMBERS_SET_SIZE];
-
-	fn type_info() -> Type {
-		Self::Identity::type_info()
+impl core::fmt::Debug for MembersSet {
+	fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+		write!(f, "MembersSet")
 	}
 }
 
-impl MaxEncodedLen for MembersSet {
-	fn max_encoded_len() -> usize {
-		MEMBERS_SET_SIZE
+impl core::cmp::PartialEq for MembersSet {
+	fn eq(&self, other: &Self) -> bool {
+		self.encode() == other.encode()
 	}
 }
+
+impl core::cmp::Eq for MembersSet {}
 
 #[derive(Clone, CanonicalDeserialize, CanonicalSerialize)]
-pub struct MembersCommitment(VerifierKey);
+pub struct MembersCommitment(bandersnatch::RingVerifierKey);
 
-ark_scale::impl_scale_via_ark!(MembersCommitment);
-
+// TODO: check
 const MEMBERS_COMMITMENT_SIZE: usize = 4 * 48 + 2 * 96; // 4 bls G1 + 2 bls G2
-
-impl scale_info::TypeInfo for MembersCommitment {
-	type Identity = [u8; MEMBERS_COMMITMENT_SIZE];
-
-	fn type_info() -> Type {
-		Self::Identity::type_info()
-	}
-}
-
-impl MaxEncodedLen for MembersCommitment {
-	fn max_encoded_len() -> usize {
-		MEMBERS_COMMITMENT_SIZE
-	}
-}
+impl_scale!(MembersCommitment, MEMBERS_COMMITMENT_SIZE);
 
 impl core::fmt::Debug for MembersCommitment {
 	fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
@@ -113,46 +93,63 @@ impl core::cmp::PartialEq for MembersCommitment {
 		self.encode() == other.encode()
 	}
 }
-
 impl core::cmp::Eq for MembersCommitment {}
 
-pub struct BandersnatchVrfVerifiable;
-
-impl BandersnatchVrfVerifiable {
-	pub fn start_members_from_params(
-		vk: KzgVk,
-		srs: impl Fn(Range<usize>) -> Result<Vec<bls12_381::G1Affine>, ()>,
-	) -> MembersSet {
-		let piop_params = bandersnatch_vrfs::ring::make_piop_params(DOMAIN_SIZE);
-		let ring = RingCommitment::empty(&piop_params, srs, vk.g1.into());
-		MembersSet {
-			ring,
-			kzg_raw_vk: vk,
-		}
-	}
-}
+const PUBLIC_KEY_LENGTH: usize = 32;
 
 #[derive(
 	Clone, Eq, PartialEq, Debug, Encode, Decode, TypeInfo, MaxEncodedLen, DecodeWithMemTracking,
 )]
 pub struct EncodedPublicKey(pub [u8; PUBLIC_KEY_LENGTH]);
 
+#[derive(Clone, Eq, PartialEq, Debug, CanonicalSerialize, CanonicalDeserialize)]
+pub struct InternalPublicKey(bandersnatch::AffinePoint);
+
+impl_scale!(InternalPublicKey, PUBLIC_KEY_LENGTH);
+
+#[derive(Clone, Eq, PartialEq, Debug, CanonicalSerialize, CanonicalDeserialize)]
+pub struct StaticChunk(ark_ec_vrfs::ring::G1Affine<bandersnatch::BandersnatchSha512Ell2>);
+
+const G1_POINT_LENGTH: usize = 48;
+impl_scale!(StaticChunk, G1_POINT_LENGTH);
+
+const IETF_SIGNATURE_SIZE: usize = 96;
+
+#[derive(CanonicalSerialize, CanonicalDeserialize)]
+struct IetfVrfSignature {
+	output: bandersnatch::Output,
+	proof: bandersnatch::IetfProof,
+}
+
+const RING_SIGNATURE_SIZE: usize = 788;
+
+#[derive(CanonicalSerialize, CanonicalDeserialize)]
+struct RingVrfSignature {
+	output: bandersnatch::Output,
+	proof: bandersnatch::RingProof,
+}
+
+#[inline(always)]
+fn make_alias(output: &bandersnatch::Output) -> Alias {
+	Alias::try_from(&output.hash()[..32]).expect("Bandersnatch suite hash is 64 bytes")
+}
+
+pub struct BandersnatchVrfVerifiable;
+
 impl GenerateVerifiable for BandersnatchVrfVerifiable {
 	type Members = MembersCommitment;
 	type Intermediate = MembersSet;
 	type Member = EncodedPublicKey;
-	type InternalMember = PublicKeyInternal;
-	type Secret = SecretKey;
-	type Commitment = (u32, ArkScale<ProverKey>);
+	type InternalMember = InternalPublicKey;
+	type Secret = bandersnatch::Secret;
+	type Commitment = (u32, ArkScale<bandersnatch::RingProverKey>);
 	type Proof = [u8; RING_SIGNATURE_SIZE];
-	type Signature = [u8; THIN_SIGNATURE_SIZE];
-	type StaticChunk = ArkScale<bls12_381::G1Affine>;
+	type Signature = [u8; IETF_SIGNATURE_SIZE];
+	type StaticChunk = StaticChunk;
 
 	fn start_members() -> Self::Intermediate {
-		MembersSet {
-			ring: EMPTY_RING,
-			kzg_raw_vk: bandersnatch_vrfs::zcash_consts::ZCASH_KZG_VK,
-		}
+		const EMPTY_RING_COMMITMENT_DATA: &[u8] = include_bytes!("ring-data/ring-builder.bin");
+		MembersSet::deserialize_uncompressed_unchecked(EMPTY_RING_COMMITMENT_DATA).unwrap()
 	}
 
 	fn push_member(
@@ -161,23 +158,25 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 		lookup: impl Fn(usize) -> Result<Self::StaticChunk, ()>,
 	) -> Result<(), ()> {
 		let who: Self::InternalMember = Self::internal_member(&who);
-		intermediate
-			.ring
-			.append(&[who.0], |range| Ok(vec![lookup(range.start)?.0]));
+		let loader = |range: Range<usize>| {
+			let item = lookup(range.start).ok()?.0;
+			Some(vec![item])
+		};
+		intermediate.0.append(&[who.0], loader);
 		Ok(())
 	}
 
-	fn finish_members(inter: Self::Intermediate) -> Self::Members {
-		let verifier_key = VerifierKey::from_ring_and_kzg_vk(&inter.ring, inter.kzg_raw_vk);
+	fn finish_members(intermediate: Self::Intermediate) -> Self::Members {
+		let verifier_key = intermediate.0.finalize();
 		MembersCommitment(verifier_key)
 	}
 
 	fn new_secret(entropy: Entropy) -> Self::Secret {
-		SecretKey::from_seed(&entropy)
+		Self::Secret::from_seed(&entropy)
 	}
 
 	fn member_from_secret(secret: &Self::Secret) -> Self::Member {
-		Self::external_member(&secret.to_public())
+		Self::external_member(&InternalPublicKey(secret.public().0))
 	}
 
 	fn validate(
@@ -186,34 +185,44 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 		context: &[u8],
 		message: &[u8],
 	) -> Result<Alias, ()> {
+		use ark_ec_vrfs::ring::Prover;
 		// This doesn't require the whole kzg. Thus is more appropriate if used on-chain
 		// Is a bit slower as it requires to recompute piop_params, but still in the order of ms
-
 		let ring_verifier =
 			bandersnatch::RingContext::verifier_no_context(members.0.clone(), RING_SIZE);
 
-		let vrf_input = Message {
-			domain: VRF_INPUT_DOMAIN,
-			message: context,
-		}
-		.into_vrf_input();
+		let input_msg = [VRF_INPUT_DOMAIN, context].concat();
+		let input = bandersnatch::Input::new(&input_msg[..]).expect("H2C can't fail here");
 
-		let ring_signature =
+		let signature =
 			RingVrfSignature::deserialize_compressed(proof.as_slice()).map_err(|_| ())?;
 
-		let ios = RingVerifier(&ring_verifier)
-			.verify_ring_vrf(message, core::iter::once(vrf_input), &ring_signature)
-			.map_err(|_| ())?;
+		bandersnatch::Public::verify(
+			input,
+			signature.output,
+			message,
+			&signature.proof,
+			&ring_verifier,
+		)
+		.map_err(|_| ())?;
 
-		let alias: Alias = ios[0].vrf_output_bytes(VRF_OUTPUT_DOMAIN);
-		Ok(alias)
+		Ok(make_alias(&signature.output))
 	}
 
+	// TODO: implement a plain Schnorr signature for Bandersnatch
 	fn sign(secret: &Self::Secret, message: &[u8]) -> Result<Self::Signature, ()> {
-		let mut transcript = Transcript::new_labeled(THIN_SIGNATURE_CONTEXT);
-		transcript.append_slice(message);
-		let signature = secret.sign_thin_vrf(transcript, &[]);
-		let mut raw = [0u8; THIN_SIGNATURE_SIZE];
+		use ark_ec_vrfs::ietf::Prover;
+		// let mut transcript = Transcript::new_labeled(THIN_SIGNATURE_CONTEXT);
+		// transcript.append_slice(message);
+		let input_msg = [THIN_SIGNATURE_CONTEXT, message].concat();
+		let input = bandersnatch::Input::new(&input_msg[..]).expect("H2C can't fail here");
+		let output = secret.output(input);
+
+		let proof = secret.prove(input, output, b"");
+
+		let signature = IetfVrfSignature { output, proof };
+
+		let mut raw = [0u8; IETF_SIGNATURE_SIZE];
 		signature
 			.serialize_compressed(raw.as_mut_slice())
 			.map_err(|_| ())?;
@@ -225,13 +234,14 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 		message: &[u8],
 		member: &Self::Member,
 	) -> bool {
-		let signature: ThinVrfSignature =
-			ThinVrfSignature::deserialize_compressed(signature.as_slice()).unwrap();
-		let mut transcript = Transcript::new_labeled(THIN_SIGNATURE_CONTEXT);
-		transcript.append_slice(message);
-		let member: Self::InternalMember = Self::internal_member(member);
-		member
-			.verify_thin_vrf(transcript, core::iter::empty::<VrfInput>(), &signature)
+		use ark_ec_vrfs::ietf::Verifier;
+		let signature = IetfVrfSignature::deserialize_compressed(signature.as_slice()).unwrap();
+		let input_msg = [THIN_SIGNATURE_CONTEXT, message].concat();
+		let input = bandersnatch::Input::new(&input_msg[..]).expect("H2C can't fail here");
+		let member = Self::internal_member(member);
+		let public = bandersnatch::Public::from(member.0);
+		public
+			.verify(input, signature.output, b"", &signature.proof)
 			.is_ok()
 	}
 
@@ -240,11 +250,13 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 		member: &Self::Member,
 		members: impl Iterator<Item = Self::Member>,
 	) -> Result<Self::Commitment, ()> {
-		let pks: Vec<_> = members.map(|m| Self::internal_member(&m).0).collect();
+		let pks = members
+			.map(|m| Self::internal_member(&m).0)
+			.collect::<Vec<_>>();
 		let member: Self::InternalMember = Self::internal_member(member);
 		let member_idx = pks.iter().position(|&m| m == member.0).ok_or(())?;
 		let member_idx = member_idx as u32;
-		let prover_key = kzg().prover_key(pks);
+		let prover_key = ring_context().prover_key(&pks[..]);
 		Ok((member_idx, prover_key.into()))
 	}
 
@@ -263,33 +275,31 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 		context: &[u8],
 		message: &[u8],
 	) -> Result<(Self::Proof, Alias), ()> {
+		use ark_ec_vrfs::ring::Prover;
 		let (prover_idx, prover_key) = commitment;
-		if prover_idx >= kzg().max_keyset_size() as u32 {
+		let ctx = ring_context();
+		if prover_idx >= ctx.max_ring_size() as u32 {
 			return Err(());
 		}
 
-		let ring_prover = kzg().init_ring_prover(prover_key.0, prover_idx as usize);
+		let ring_prover = ctx.prover(prover_key.0, prover_idx as usize);
 
-		let vrf_input = Message {
-			domain: VRF_INPUT_DOMAIN,
-			message: context,
-		}
-		.into_vrf_input();
+		let input_msg = [VRF_INPUT_DOMAIN, context].concat();
+		let input = bandersnatch::Input::new(&input_msg[..]).expect("H2C can't fail here");
+		let preout = secret.output(input);
+		let alias = make_alias(&preout);
 
-		let ios = [secret.vrf_inout(vrf_input)];
+		let proof = secret.prove(input, preout, message, &ring_prover);
 
-		let signature: RingVrfSignature = RingProver {
-			ring_prover: &ring_prover,
-			secret,
-		}
-		.sign_ring_vrf(message, &ios);
+		let signature = RingVrfSignature {
+			output: preout,
+			proof,
+		};
 
 		let mut buf = [0u8; RING_SIGNATURE_SIZE];
 		signature
 			.serialize_compressed(buf.as_mut_slice())
 			.map_err(|_| ())?;
-
-		let alias: Alias = ios[0].vrf_output_bytes(VRF_OUTPUT_DOMAIN);
 
 		Ok((buf, alias))
 	}
@@ -305,15 +315,10 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 	}
 
 	fn alias_in_context(secret: &Self::Secret, context: &[u8]) -> Result<Alias, ()> {
-		let vrf_input = Message {
-			domain: VRF_INPUT_DOMAIN,
-			message: context,
-		}
-		.into_vrf_input();
-
-		let ios = [secret.vrf_inout(vrf_input)];
-
-		let alias: Alias = ios[0].vrf_output_bytes(VRF_OUTPUT_DOMAIN);
+		let input_msg = [VRF_INPUT_DOMAIN, context].concat();
+		let input = bandersnatch::Input::new(&input_msg[..]).expect("H2C can't fail here");
+		let output = secret.output(input);
+		let alias = make_alias(&output);
 		Ok(alias)
 	}
 
@@ -326,20 +331,46 @@ impl GenerateVerifiable for BandersnatchVrfVerifiable {
 	}
 
 	fn internal_member(value: &Self::Member) -> Self::InternalMember {
-		deserialize_publickey(&value.0[..]).expect("must be valid")
+		let pt =
+			bandersnatch::AffinePoint::deserialize_compressed(&value.0[..]).expect("must be valid");
+		InternalPublicKey(pt.into())
 	}
 }
 
 #[cfg(test)]
 mod tests {
-	use bandersnatch_vrfs::ring::StaticVerifierKey;
-
 	use super::*;
+	use ark_ec_vrfs::{ring::SrsLookup, suites::bandersnatch::BandersnatchSha512Ell2};
 
-	#[cfg(feature = "small-ring")]
-	const ONCHAIN_VK: &[u8] = include_bytes!("ring-data/zcash-9.vk");
-	#[cfg(not(feature = "small-ring"))]
-	const ONCHAIN_VK: &[u8] = include_bytes!("ring-data/zcash-16.vk");
+	type RingBuilderPcsParams = ark_ec_vrfs::ring::RingBuilderPcsParams<BandersnatchSha512Ell2>;
+
+	impl BandersnatchVrfVerifiable {
+		fn start_members_from_params() -> (MembersSet, RingBuilderPcsParams) {
+			let (builder, builder_pcs_params) = ring_context().verifier_key_builder();
+			(MembersSet(builder), builder_pcs_params)
+		}
+	}
+
+	// TODO: skip
+	#[test]
+	fn generate_empty_ring_builder() {
+		use std::io::Write;
+		const EMPTY_RING_COMMITMENT_FILE: &str = concat!(
+			env!("CARGO_MANIFEST_DIR"),
+			"/src/ring-data/ring-builder.bin"
+		);
+		let (builder, builder_pcs_params) = BandersnatchVrfVerifiable::start_members_from_params();
+		let mut buf = Vec::with_capacity(builder.uncompressed_size());
+		builder.serialize_uncompressed(&mut buf).unwrap();
+		println!("Writing empty ring to: {}", EMPTY_RING_COMMITMENT_FILE);
+		let mut file = std::fs::File::create(EMPTY_RING_COMMITMENT_FILE).unwrap(); // Create or truncate the file
+		file.write_all(&buf).unwrap();
+	}
+
+	#[test]
+	fn check_backend_assumptions() {
+		todo!("Check serialized sizes");
+	}
 
 	#[test]
 	fn start_push_finish() {
@@ -351,13 +382,16 @@ mod tests {
 		let bob = BandersnatchVrfVerifiable::member_from_secret(&bob_sec);
 		let charlie = BandersnatchVrfVerifiable::member_from_secret(&charlie_sec);
 
-		let vk = StaticVerifierKey::deserialize_uncompressed_unchecked(ONCHAIN_VK).unwrap();
-		let get_one = |i| Ok(ArkScale(vk.lag_g1[i]));
-		let get_many = |range: Range<usize>| Ok(vk.lag_g1[range].to_vec());
-
 		let mut inter1 = BandersnatchVrfVerifiable::start_members();
-		let mut inter2 = BandersnatchVrfVerifiable::start_members_from_params(vk.kzg_vk, get_many);
+		let (mut inter2, builder_params) = BandersnatchVrfVerifiable::start_members_from_params();
 		assert_eq!(inter1, inter2);
+
+		let get_one = |i| {
+			(&builder_params)
+				.lookup(i..i + 1)
+				.map(|v| StaticChunk(v[0]))
+				.ok_or(())
+		};
 
 		BandersnatchVrfVerifiable::push_member(&mut inter1, alice.clone(), get_one).unwrap();
 		BandersnatchVrfVerifiable::push_member(&mut inter2, alice.clone(), get_one).unwrap();
@@ -390,8 +424,11 @@ mod tests {
 		let message = b"FooBar";
 
 		let start = Instant::now();
-		let _ = kzg();
-		println!("* KZG decode: {} ms", (Instant::now() - start).as_millis());
+		let _ = ring_context();
+		println!(
+			"* PCS params decode: {} ms",
+			(Instant::now() - start).as_millis()
+		);
 
 		let members: Vec<_> = (0..10)
 			.map(|i| {
@@ -414,8 +451,14 @@ mod tests {
 		println!("* Create: {} ms", (Instant::now() - start).as_millis());
 		println!("  Proof size: {} bytes", proof.encode().len()); // 788 bytes
 
-		let vk = StaticVerifierKey::deserialize_uncompressed_unchecked(ONCHAIN_VK).unwrap();
-		let get_one = |i| Ok(ArkScale(vk.lag_g1[i]));
+		let (mut builder, builder_params) = BandersnatchVrfVerifiable::start_members_from_params();
+
+		let get_one = |i| {
+			(&builder_params)
+				.lookup(i..i + 1)
+				.map(|v| StaticChunk(v[0]))
+				.ok_or(())
+		};
 
 		let start = Instant::now();
 		let mut inter = BandersnatchVrfVerifiable::start_members();
